+++ packages/ranger-utils/src/index.ts
@@ -1,2 +1,3 @@
 export { parseCookies, destroyCookie, setItem, getItem, removeItem } from './cookies'
+export type { Options } from './cookies'
 export { events } from './events'