+++ apps/ranger-vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ranger-theme/vite
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ranger-theme/hooks@0.3.1
+
 ## 0.1.4
 
 ### Patch Changes