+++ packages/ranger-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ranger-theme/utils
 
+## 0.5.1
+
+### Patch Changes
+
+- cookie params for ssr context null
+
 ## 0.5.0
 
 ### Minor Changes